api/requirements.txt
Flask
jsonify
flask_sqlalchemy
flask-restful
Flask-Login==0.4.1
Flask-WTF==0.14.2
Flask-Migrate==2.5.2
pymysql
mysqlclient
SQLAlchemy==1.3.1
python-slugify
apscheduler
Flask-Scss
Flask-Bootstrap4
email_validator
psycopg2
uwsgi
gunicorn
wheel
Flask-Caching
Flask-Mail==0.9.1
pytest
werkzeug==0.16.0
requests
Flask-FontAwesome==0.1.5
Flask-Babel
stripe>=2.48.0,<3.0
